backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.4
alphashape==1.3.1
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
asyncer==0.0.8
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.9.0
Brotli==1.1.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
click-log==0.4.0
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.3
Crawl4AI==0.7.4
cryptography==45.0.6
cycler==0.12.1
datasets==4.0.0
decorator==5.2.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dspy==3.0.2
executing==2.2.0
fake-http-header==0.3.5
fake-useragent==2.2.0
Faker==37.5.3
filelock==3.19.1
Flask==3.1.1
flask-cors==6.0.1
Flask-SocketIO==5.5.1
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2025.3.0
gepa==0.0.4
gevent==25.5.1
gevent-websocket==0.10.1
gower==0.1.2
greenlet==3.2.4
h11==0.16.0
h2==4.2.0
hf-xet==1.1.7
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json_repair==0.50.0
jsonpickle==4.1.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lark==1.2.2
litellm==1.75.8
lxml==5.4.0
magicattr==0.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.6.4
multiprocess==0.70.16
networkx==3.5
nltk==3.9.1
numpy==2.3.2
ollama==0.5.3
openai==1.99.9
optuna==4.5.0
packaging==25.0
pandas==2.3.1
parso==0.8.5
patchright==1.52.5
pexpect==4.9.0
pillow==11.3.0
playwright==1.54.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyMySQL==1.1.2
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
rake-nltk==1.0.6
rank-bm25==0.2.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rpds-py==0.27.0
rtree==1.4.1
scikit-learn==1.7.1
scipy==1.16.1
shapely==2.1.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
tenacity==9.1.2
tf-playwright-stealth==1.2.0
threadpoolctl==3.6.0
tidb-vector==0.0.15
tiktoken==0.11.0
tokenizers==0.21.4
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.7.4
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wordcloud==1.9.4
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zope.event==5.1.1
zope.interface==7.2
